game demo for BeMyApp livecode Android #2
You are a lone adventurer in search for a cursed treasure, about to
enter a mysterious temple on the top of a mountain, almost invisible
due to permanent snow storm...

This is a game prototype I developed for this meetup:
http://www.meetup.com/Paris-Android-LiveCode/events/102301792/

It was destined to demonstrate Corona SDK graphic capabilities and how
they can help you to to create an immersive atmosphere (with the help
of good audio design).

We also used this prototype to demonstrate Intel's GPA: a profiling
tool for x86 chips on smartphones. More info here:
http://software.intel.com/en-us/vcsource/tools/intel-gpa